+++ CHANGELOG.md
@@ -6,8 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v1.6.0]
+
 ### Added
-- This week filter for database queries.
+- This week on date filter for database queries. (#184)
+- Relation filter for database queries. (#185) 
 
 ## [1.5.0]
 
@@ -285,4 +288,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [1.4.0]: https://github.com/mariosimao/notion-sdk-php/releases/tag/v1.4.0
 [1.4.1]: https://github.com/mariosimao/notion-sdk-php/releases/tag/v1.4.1
 [1.5.0]: https://github.com/mariosimao/notion-sdk-php/releases/tag/v1.5.0
-[Unreleased]: https://github.com/mariosimao/notion-sdk-php/compare/v1.5.0...HEAD
+[1.6.0]: https://github.com/mariosimao/notion-sdk-php/releases/tag/v1.6.0
+[Unreleased]: https://github.com/mariosimao/notion-sdk-php/compare/v1.6.0...HEAD